# 🛡️ feat: Upgrade Security System to Match Self-Healing Sophistication