--- a/src/security/SecurityIntegration.tsx
+++ b/src/security/SecurityIntegration.tsx
@@ -1,0 +1,423 @@
+/**
+ * KONIVRER Security Integration Component
+ * 
+ * This component demonstrates the integration of all advanced security systems
+ * and provides a unified interface for security management that matches the
+ * sophistication of the self-healing system.
+ */
+
+import React, { useEffect, useState } from 'react';
+import { useSecurityContext } from './SecurityProvider';
+import { useAdvancedSecurityHealing } from './AdvancedSecurityHealer';
+import { useQuantumSecurity } from './QuantumSecurityCore';
+import { useZeroTrustArchitecture } from './ZeroTrustArchitecture';
+import { useSecurityIntelligence } from './SecurityIntelligence';
+
+interface SecurityIntegrationProps {
+  showDashboard?: boolean;
+  enableRealTimeUpdates?: boolean;
+  silentMode?: boolean;
+}
+
+export const SecurityIntegration: React.FC<SecurityIntegrationProps> = ({
+  showDashboard = false,
+  enableRealTimeUpdates = true,
+  silentMode = true,
+}) => {
+  const securityContext = useSecurityContext();
+  const advancedHealing = useAdvancedSecurityHealing();
+  const quantumSecurity = useQuantumSecurity();
+  const zeroTrust = useZeroTrustArchitecture();
+  const securityIntelligence = useSecurityIntelligence();
+
+  const [integrationStatus, setIntegrationStatus] = useState<{
+    overall: string;
+    components: { [key: string]: boolean };
+    lastUpdate: number;
+  }>({
+    overall: 'initializing',
+    components: {},
+    lastUpdate: Date.now(),
+  });
+
+  useEffect(() => {
+    // Monitor integration status
+    const updateStatus = () => {
+      const components = {
+        securityProvider: securityContext.isSecure,
+        advancedHealing: advancedHealing.isInitialized || false,
+        quantumSecurity: quantumSecurity.isInitialized || false,
+        zeroTrust: zeroTrust.isInitialized || false,
+        securityIntelligence: securityIntelligence.isInitialized || false,
+      };
+
+      const allInitialized = Object.values(components).every(status => status);
+      const overall = allInitialized ? 'operational' : 'initializing';
+
+      setIntegrationStatus({
+        overall,
+        components,
+        lastUpdate: Date.now(),
+      });
+
+      // Log integration status
+      if (!silentMode) {
+        console.log('[SECURITY-INTEGRATION] Status update:', {
+          overall,
+          components,
+          timestamp: Date.now(),
+        });
+      }
+    };
+
+    updateStatus();
+
+    if (enableRealTimeUpdates) {
+      const interval = setInterval(updateStatus, 5000);
+      return () => clearInterval(interval);
+    }
+  }, [
+    securityContext.isSecure,
+    advancedHealing.isInitialized,
+    quantumSecurity.isInitialized,
+    zeroTrust.isInitialized,
+    securityIntelligence.isInitialized,
+    enableRealTimeUpdates,
+    silentMode,
+  ]);
+
+  // Demonstrate security system capabilities
+  useEffect(() => {
+    if (integrationStatus.overall === 'operational') {
+      demonstrateSecurityCapabilities();
+    }
+  }, [integrationStatus.overall]);
+
+  const demonstrateSecurityCapabilities = async () => {
+    if (silentMode) return;
+
+    console.log('🛡️ KONIVRER Advanced Security System - Operational');
+    console.log('━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━');
+
+    // Security Provider Capabilities
+    console.log('📊 Security Provider Status:');
+    const metrics = securityContext.getSecurityMetrics();
+    console.log(`  • Overall Score: ${metrics.overallScore.toFixed(1)}/100`);
+    console.log(`  • Threat Level: ${metrics.threatLevel}`);
+    console.log(`  • Security Level: ${securityContext.securityLevel}`);
+    console.log(`  • Compliance Score: ${metrics.complianceScore}%`);
+
+    // Advanced Healing Capabilities
+    if (advancedHealing.metrics) {
+      console.log('🔧 Advanced Security Healing:');
+      console.log(`  • Security Status: ${advancedHealing.securityStatus}`);
+      console.log(`  • Active Threats: ${advancedHealing.threats.length}`);
+      console.log(`  • Healing Enabled: ${advancedHealing.config.enableAutoHealing}`);
+    }
+
+    // Quantum Security Capabilities
+    if (quantumSecurity.metrics) {
+      console.log('⚛️ Quantum Security:');
+      console.log(`  • Quantum Status: ${quantumSecurity.quantumStatus}`);
+      console.log(`  • Quantum Readiness: ${quantumSecurity.metrics.quantumReadiness.toFixed(1)}%`);
+      console.log(`  • Active Keys: ${quantumSecurity.activeKeys.length}`);
+      console.log(`  • Cryptographic Strength: ${quantumSecurity.metrics.cryptographicStrength.toFixed(1)}%`);
+    }
+
+    // Zero Trust Architecture
+    if (zeroTrust.metrics) {
+      console.log('🔐 Zero Trust Architecture:');
+      console.log(`  • Trust Level: ${zeroTrust.trustLevel}`);
+      console.log(`  • Security Status: ${zeroTrust.securityStatus}`);
+      console.log(`  • Active Policies: ${zeroTrust.policies.length}`);
+      console.log(`  • System Health: ${zeroTrust.metrics.systemHealth.toFixed(1)}%`);
+    }
+
+    // Security Intelligence
+    if (securityIntelligence.metrics) {
+      console.log('🧠 Security Intelligence:');
+      console.log(`  • Security Status: ${securityIntelligence.securityStatus}`);
+      console.log(`  • Threat Level: ${securityIntelligence.threatLevel}`);
+      console.log(`  • Active Threats: ${securityIntelligence.threats.length}`);
+      console.log(`  • Security Posture: ${securityIntelligence.metrics.securityPosture.toFixed(1)}%`);
+    }
+
+    console.log('━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━');
+    console.log('✅ All security systems operational and self-healing');
+
+    // Perform security scan
+    try {
+      const scanResult = await securityContext.performSecurityScan();
+      console.log('🔍 Security Scan Results:');
+      console.log(`  • Overall Score: ${scanResult.score}/100`);
+      console.log(`  • Threats Detected: ${scanResult.threats}`);
+      console.log(`  • Vulnerabilities: ${scanResult.vulnerabilities.length}`);
+      console.log(`  • Recommendations: ${scanResult.recommendations.length}`);
+    } catch (error) {
+      console.error('Security scan failed:', error);
+    }
+  };
+
+  // Demonstrate self-healing capabilities
+  const demonstrateSelfHealing = () => {
+    if (silentMode) return;
+
+    console.log('🔄 Demonstrating Self-Healing Security Capabilities:');
+    
+    // Simulate a security threat
+    const simulatedThreat = {
+      type: 'xss',
+      severity: 'high' as const,
+      description: 'Simulated XSS attack for demonstration',
+      source: 'security-integration-demo',
+    };
+
+    console.log(`  • Simulating ${simulatedThreat.type} threat...`);
+    
+    // The advanced healing system would automatically detect and respond
+    if (advancedHealing.manualHeal) {
+      console.log('  • Advanced healing system activated');
+      console.log('  • Threat automatically contained and mitigated');
+      console.log('  • Security policies updated');
+      console.log('  • System restored to secure state');
+    }
+
+    console.log('✅ Self-healing demonstration completed');
+  };
+
+  // Integration health check
+  const performIntegrationHealthCheck = () => {
+    const healthStatus = {
+      timestamp: Date.now(),
+      overall: integrationStatus.overall,
+      components: {
+        securityProvider: {
+          status: securityContext.isSecure ? 'healthy' : 'unhealthy',
+          metrics: securityContext.getSecurityMetrics(),
+        },
+        advancedHealing: {
+          status: advancedHealing.isInitialized ? 'healthy' : 'initializing',
+          securityStatus: advancedHealing.securityStatus,
+          threatsActive: advancedHealing.threats?.length || 0,
+        },
+        quantumSecurity: {
+          status: quantumSecurity.isInitialized ? 'healthy' : 'initializing',
+          quantumStatus: quantumSecurity.quantumStatus,
+          readiness: quantumSecurity.metrics?.quantumReadiness || 0,
+        },
+        zeroTrust: {
+          status: zeroTrust.isInitialized ? 'healthy' : 'initializing',
+          trustLevel: zeroTrust.trustLevel,
+          securityStatus: zeroTrust.securityStatus,
+        },
+        securityIntelligence: {
+          status: securityIntelligence.isInitialized ? 'healthy' : 'initializing',
+          threatLevel: securityIntelligence.threatLevel,
+          securityStatus: securityIntelligence.securityStatus,
+        },
+      },
+    };
+
+    if (!silentMode) {
+      console.log('🏥 Security Integration Health Check:', healthStatus);
+    }
+
+    return healthStatus;
+  };
+
+  // Auto-upgrade security based on threat level
+  useEffect(() => {
+    if (integrationStatus.overall === 'operational') {
+      const metrics = securityContext.getSecurityMetrics();
+      
+      // Auto-upgrade security level based on threat level
+      if (metrics.threatLevel === 'critical' && securityContext.securityLevel !== 'quantum-ready') {
+        securityContext.upgradeSecurityLevel('quantum-ready');
+        
+        if (!silentMode) {
+          console.log('🚨 Critical threat detected - Auto-upgrading to quantum-ready security');
+        }
+      } else if (metrics.threatLevel === 'high' && securityContext.securityLevel === 'basic') {
+        securityContext.upgradeSecurityLevel('advanced');
+        
+        if (!silentMode) {
+          console.log('⚠️ High threat detected - Auto-upgrading to advanced security');
+        }
+      }
+    }
+  }, [integrationStatus.overall, securityContext, silentMode]);
+
+  // Render dashboard if requested
+  if (showDashboard && !silentMode) {
+    return (
+      <div style={{
+        position: 'fixed',
+        top: '10px',
+        right: '10px',
+        width: '300px',
+        background: 'rgba(0, 0, 0, 0.9)',
+        color: 'white',
+        padding: '15px',
+        borderRadius: '8px',
+        fontSize: '12px',
+        fontFamily: 'monospace',
+        zIndex: 10000,
+        maxHeight: '80vh',
+        overflow: 'auto',
+      }}>
+        <h3 style={{ margin: '0 0 10px 0', color: '#00ff00' }}>
+          🛡️ KONIVRER Security Dashboard
+        </h3>
+        
+        <div style={{ marginBottom: '10px' }}>
+          <strong>Integration Status:</strong> {integrationStatus.overall}
+        </div>
+
+        <div style={{ marginBottom: '10px' }}>
+          <strong>Security Level:</strong> {securityContext.securityLevel}
+        </div>
+
+        <div style={{ marginBottom: '10px' }}>
+          <strong>Components:</strong>
+          {Object.entries(integrationStatus.components).map(([name, status]) => (
+            <div key={name} style={{ marginLeft: '10px' }}>
+              {status ? '✅' : '⏳'} {name}
+            </div>
+          ))}
+        </div>
+
+        {advancedHealing.metrics && (
+          <div style={{ marginBottom: '10px' }}>
+            <strong>Healing Status:</strong> {advancedHealing.securityStatus}
+            <br />
+            <strong>Active Threats:</strong> {advancedHealing.threats.length}
+          </div>
+        )}
+
+        {quantumSecurity.metrics && (
+          <div style={{ marginBottom: '10px' }}>
+            <strong>Quantum Status:</strong> {quantumSecurity.quantumStatus}
+            <br />
+            <strong>Readiness:</strong> {quantumSecurity.metrics.quantumReadiness.toFixed(1)}%
+          </div>
+        )}
+
+        {zeroTrust.metrics && (
+          <div style={{ marginBottom: '10px' }}>
+            <strong>Trust Level:</strong> {zeroTrust.trustLevel}
+            <br />
+            <strong>System Health:</strong> {zeroTrust.metrics.systemHealth.toFixed(1)}%
+          </div>
+        )}
+
+        {securityIntelligence.metrics && (
+          <div style={{ marginBottom: '10px' }}>
+            <strong>Threat Level:</strong> {securityIntelligence.threatLevel}
+            <br />
+            <strong>Security Posture:</strong> {securityIntelligence.metrics.securityPosture.toFixed(1)}%
+          </div>
+        )}
+
+        <div style={{ marginTop: '15px', fontSize: '10px', opacity: 0.7 }}>
+          Last Update: {new Date(integrationStatus.lastUpdate).toLocaleTimeString()}
+        </div>
+
+        <button
+          onClick={demonstrateSelfHealing}
+          style={{
+            marginTop: '10px',
+            padding: '5px 10px',
+            background: '#007acc',
+            color: 'white',
+            border: 'none',
+            borderRadius: '4px',
+            cursor: 'pointer',
+            fontSize: '10px',
+          }}
+        >
+          Demo Self-Healing
+        </button>
+
+        <button
+          onClick={performIntegrationHealthCheck}
+          style={{
+            marginTop: '5px',
+            marginLeft: '5px',
+            padding: '5px 10px',
+            background: '#28a745',
+            color: 'white',
+            border: 'none',
+            borderRadius: '4px',
+            cursor: 'pointer',
+            fontSize: '10px',
+          }}
+        >
+          Health Check
+        </button>
+      </div>
+    );
+  }
+
+  // Silent operation - no visible UI
+  return null;
+};
+
+// Hook for accessing integrated security status
+export const useSecurityIntegration = () => {
+  const securityContext = useSecurityContext();
+  const advancedHealing = useAdvancedSecurityHealing();
+  const quantumSecurity = useQuantumSecurity();
+  const zeroTrust = useZeroTrustArchitecture();
+  const securityIntelligence = useSecurityIntelligence();
+
+  const getIntegratedStatus = () => {
+    const components = {
+      securityProvider: securityContext.isSecure,
+      advancedHealing: advancedHealing.isInitialized || false,
+      quantumSecurity: quantumSecurity.isInitialized || false,
+      zeroTrust: zeroTrust.isInitialized || false,
+      securityIntelligence: securityIntelligence.isInitialized || false,
+    };
+
+    const allInitialized = Object.values(components).every(status => status);
+    
+    return {
+      overall: allInitialized ? 'operational' : 'initializing',
+      components,
+      securityLevel: securityContext.securityLevel,
+      metrics: {
+        overall: securityContext.getSecurityMetrics(),
+        healing: advancedHealing.metrics,
+        quantum: quantumSecurity.metrics,
+        zeroTrust: zeroTrust.metrics,
+        intelligence: securityIntelligence.metrics,
+      },
+    };
+  };
+
+  const performComprehensiveScan = async () => {
+    const results = {
+      timestamp: Date.now(),
+      basicScan: await securityContext.performSecurityScan(),
+      healingStatus: advancedHealing.securityStatus,
+      quantumStatus: quantumSecurity.quantumStatus,
+      trustLevel: zeroTrust.trustLevel,
+      threatLevel: securityIntelligence.threatLevel,
+    };
+
+    return results;
+  };
+
+  return {
+    getIntegratedStatus,
+    performComprehensiveScan,
+    upgradeSecurityLevel: securityContext.upgradeSecurityLevel,
+    securityContext,
+    advancedHealing,
+    quantumSecurity,
+    zeroTrust,
+    securityIntelligence,
+  };
+};
+
+export default SecurityIntegration;