… AUTONOMOUS_SECURITY.md
… SECURITY_UPGRADE_SUMMARY.md
… src/security/AdvancedSecurityHealer.tsx
… src/security/QuantumSecurityCore.tsx
… src/security/SecurityIntelligence.tsx
… src/security/ZeroTrustArchitecture.tsx